uses .NET framework IReadOnlyList. moves terminals to subfolder
