+++ libraries/Pliant/AndNode.cs
@@ -1,8 +1,5 @@
-﻿using System;
+﻿using Pliant.Collections;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Pliant
 {
