+++ libraries/Pliant/Collections/ReadOnlyList.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using System.Collections.ObjectModel;
+
+namespace Pliant.Collections
+{
+    public class ReadOnlyList<T> : ReadOnlyCollection<T>,  IReadOnlyList<T>
+    {
+        public ReadOnlyList(IList<T> list)
+            : base(list)
+        {            
+        }
+    }
+}
