+++ libraries/Pliant/IGrammar.cs
@@ -1,5 +1,6 @@
-﻿using System;
+﻿using Pliant.Collections;
 using System.Collections.Generic;
+
 namespace Pliant
 {
     public interface IGrammar
