--- libraries/Pliant/IReadOnlyList.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Pliant
-{
-    public interface IReadOnlyList<T> : IReadOnlyCollection<T>
-    {
-        T this[int index] { get; }
-    }
-}
