--- libraries/Pliant/ITerminal.cs
@@ -1,8 +1,0 @@
-﻿using System;
-namespace Pliant
-{
-    public interface ITerminal : ISymbol
-    {
-        bool IsMatch(char character);
-    }
-}
