+++ libraries/Pliant/IntermediateNode.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using Pliant.Collections;
 
 namespace Pliant
 {
