+++ libraries/Pliant/Pliant.csproj
@@ -40,12 +40,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AndNode.cs" />
-    <Compile Include="AnyTerminal.cs" />
+    <Compile Include="Terminals\AnyTerminal.cs" />
     <Compile Include="Assert.cs" />
-    <Compile Include="CharacterClassTerminal.cs" />
+    <Compile Include="Terminals\CharacterClassTerminal.cs" />
     <Compile Include="Chart.cs" />
     <Compile Include="Collections\IEnumerableExtensions.cs" />
-    <Compile Include="DigitTerminal.cs" />
+    <Compile Include="Terminals\DigitTerminal.cs" />
     <Compile Include="DottedRule.cs" />
     <Compile Include="EarleySet.cs" />
     <Compile Include="Grammar.cs" />
@@ -69,12 +69,11 @@
     <Compile Include="IParseInterface.cs" />
     <Compile Include="IProduction.cs" />
     <Compile Include="IProductionBuilder.cs" />
-    <Compile Include="IReadOnlyList.cs" />
     <Compile Include="IRuleBuilder.cs" />
     <Compile Include="IState.cs" />
     <Compile Include="ISymbol.cs" />
     <Compile Include="ISymbolNode.cs" />
-    <Compile Include="ITerminal.cs" />
+    <Compile Include="Terminals\ITerminal.cs" />
     <Compile Include="ITerminalBuilder.cs" />
     <Compile Include="ITerminalNode.cs" />
     <Compile Include="IToken.cs" />
@@ -91,11 +90,11 @@
     <Compile Include="ParseEngine.cs" />
     <Compile Include="ParseInterface.cs" />
     <Compile Include="ParseRunner.cs" />
-    <Compile Include="ReadWriteList.cs" />
+    <Compile Include="Collections\ReadWriteList.cs" />
     <Compile Include="RegexGrammar.cs" />
     <Compile Include="RegexParser.cs" />
-    <Compile Include="SetTerminal.cs" />
-    <Compile Include="NegationTerminal.cs" />
+    <Compile Include="Terminals\SetTerminal.cs" />
+    <Compile Include="Terminals\NegationTerminal.cs" />
     <Compile Include="NonTerminal.cs" />
     <Compile Include="Observable.cs" />
     <Compile Include="Observer.cs" />
@@ -103,11 +102,11 @@
     <Compile Include="Production.cs" />
     <Compile Include="ProductionBuilder.cs" />
     <Compile Include="PulseRecognizer.cs" />
-    <Compile Include="RangeTerminal.cs" />
+    <Compile Include="Terminals\RangeTerminal.cs" />
     <Compile Include="Repetition.cs" />
     <Compile Include="RuleBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ReadOnlyList.cs" />
+    <Compile Include="Collections\ReadOnlyList.cs" />
     <Compile Include="Recognizer.cs" />
     <Compile Include="State.cs" />
     <Compile Include="StateQueue.cs" />
@@ -115,7 +114,7 @@
     <Compile Include="SymbolNode.cs" />
     <Compile Include="SymbolType.cs" />
     <Compile Include="Symbol.cs" />
-    <Compile Include="Terminal.cs" />
+    <Compile Include="Terminals\Terminal.cs" />
     <Compile Include="TerminalBuilder.cs" />
     <Compile Include="TerminalLexeme.cs" />
     <Compile Include="TerminalLexerRule.cs" />
@@ -125,7 +124,10 @@
     <Compile Include="TokenType.cs" />
     <Compile Include="TransitionState.cs" />
     <Compile Include="VirtualNode.cs" />
-    <Compile Include="WhitespaceTerminal.cs" />
+    <Compile Include="Terminals\WhitespaceTerminal.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Charts\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
