--- libraries/Pliant/ReadWriteList.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Pliant
-{
-    public class ReadWriteList<T> : List<T>, IList<T>, IReadOnlyList<T>
-    {        
-    }
-}
