+++ libraries/Pliant/RegexGrammar.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Pliant.Terminals;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
