+++ libraries/Pliant/Terminals/ITerminal.cs
@@ -1,0 +1,8 @@
+﻿using System;
+namespace Pliant.Terminals
+{
+    public interface ITerminal : ISymbol
+    {
+        bool IsMatch(char character);
+    }
+}
