+++ tests/Pliant.Tests.Unit/AnyTerminalTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Pliant.Terminals;
 
 namespace Pliant.Tests.Unit
 {
