+++ tests/Pliant.Tests.Unit/BnfTests.cs
@@ -3,6 +3,7 @@ using System.Text;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.IO;
+using Pliant.Terminals;
 
 namespace Pliant.Tests.Unit
 {
