+++ tests/Pliant.Tests.Unit/ChartTests.cs
@@ -2,6 +2,7 @@
 using System.Text;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Pliant.Terminals;
 
 namespace Pliant.Tests.Unit
 {
