+++ tests/Pliant.Tests.Unit/LexemeTests.cs
@@ -2,6 +2,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
 using System.Linq;
+using Pliant.Terminals;
 
 namespace Pliant.Tests.Unit
 {
