+++ tests/Pliant.Tests.Unit/PulseRecognizerTests.cs
@@ -3,6 +3,7 @@ using System.Text;
 using System.Collections.Generic;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.IO;
+using Pliant.Terminals;
 
 namespace Pliant.Tests.Unit
 {
