+++ tests/Pliant.Tests.Unit/TerminalLexemeTests.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Pliant.Terminals;
 using System;
 using System.Collections.Generic;
 using System.Linq;
